pyproject.toml
[tool.poetry]
name = "Publ"
version = "0.7.14"
description = "A flexible web-based publishing framework"
authors = ["fluffy <fluffy@beesbuzz.biz>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/PlaidWeb/Publ"
homepage = "https://publ.plaidweb.site/"

[tool.poetry.dependencies]
python = "^3.7"
authl = { version = "^0.6.0", optional = true }
misaka = "^2.1.1"
arrow = "^0.17.0"
atomicwrites = "^1.4.0"
awesome-slugify = "^1.6.5"
Flask = "^2.0.0"
Flask-Caching = "^1.9.0"
pony = "^0.7.14"
Pygments = "^2.12.0"
watchdog = "^1.0.2"
Whoosh = { version = "^2.7.4", optional = true }
Pillow = "^9.0.0"
pillow = "^9.3.0"
flask = "^2.2.2"

[tool.poetry.extras]
auth = ["authl"]
search = ["Whoosh"]

[tool.poetry.dev-dependencies]
autopep8 = "*"
coverage = "*"
flake8 = "*"
mypy = "*"
pylint = "*"
pytest = "*"
isort = "^5"
pytest-mock = "^3.2.0"
requests-mock = "^1.9.3"
types-python-dateutil = "^2.8.10"
types-requests = "^2.27.13"
types-python-slugify = "^5.0.3"
types-atomicwrites = "^1.4.3"
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
